feat: remove Prisma from examples/blog (login port + drizzle seed)
Port login to db.query; replace prisma/seed.js with db/seed.server.ts
(insert + onConflictDoNothing, sqlite/pg portable, run via webjs db seed).
Remove @prisma/client + prisma deps, lib/prisma.server.ts, prisma/ (schema,
seed, migrations). package.json: db:seed -> webjs db seed; drop the
prisma generate dev task (drizzle has no codegen); start.before ->
webjs db migrate. DATABASE_URL + .gitignore point at db/dev.db.
Blog test files still reference prisma; ported next. Refs #551 #563.
