+++ examples/blog/.env.example
@@ -3,7 +3,7 @@
 
 PORT=5004
 AUTH_SECRET=change-me-to-a-random-string-at-least-32-chars
-DATABASE_URL=file:./dev.db
+DATABASE_URL=file:./db/dev.db
 
 # Uncomment for OAuth login:
 # AUTH_GOOGLE_ID=
