--- examples/blog/prisma/schema.prisma
@@ -1,55 +1,0 @@
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  provider = "sqlite"
-  url      = "file:./dev.db"
-}
-
-model User {
-  id           Int       @id @default(autoincrement())
-  email        String    @unique
-  passwordHash String
-  name         String?
-  createdAt    DateTime  @default(now())
-  posts        Post[]
-  comments     Comment[]
-  sessions     Session[]
-}
-
-model Session {
-  token     String   @id
-  userId    Int
-  user      User     @relation(fields: [userId], references: [id], onDelete: Cascade)
-  expiresAt DateTime
-  createdAt DateTime @default(now())
-
-  @@index([userId])
-}
-
-model Post {
-  id        Int       @id @default(autoincrement())
-  slug      String    @unique
-  title     String
-  body      String
-  authorId  Int
-  author    User      @relation(fields: [authorId], references: [id], onDelete: Cascade)
-  createdAt DateTime  @default(now())
-  comments  Comment[]
-
-  @@index([authorId])
-  @@index([createdAt])
-}
-
-model Comment {
-  id        Int      @id @default(autoincrement())
-  postId    Int
-  post      Post     @relation(fields: [postId], references: [id], onDelete: Cascade)
-  authorId  Int
-  author    User     @relation(fields: [authorId], references: [id], onDelete: Cascade)
-  body      String
-  createdAt DateTime @default(now())
-
-  @@index([postId, createdAt])
-}
